Small Bug Fixes
Changed int to long long wherever required. Handled edge case in preprocess function.

+++ src/ApproxDistanceOracles.h
@@ -19,26 +19,28 @@ class ApproxDistanceOracles {
     bool spaceopt;  // Rerun for Space Optimisation Enablement.
     bool debug;     // For Custom Landmarks
 
-    std::vector<std::vector<int>> landmarks;
-    std::vector<int> rank;
-    std::vector<std::vector<int>> focus;
+    std::vector<std::vector<long long>> landmarks;
+    std::vector<long long> rank;
+    std::vector<std::vector<long long>> focus;
     std::vector<std::vector<double>> focus_distance;
-    std::vector<std::vector<std::queue<std::pair<int, double>>>>
+    std::vector<std::vector<std::queue<std::pair<long long, double>>>>
         ball;  // Level-wise
-    std::vector<fph::MetaFphMap<int, double, fph::meta::MixSeedHash<int>>>
+    std::vector<
+        fph::MetaFphMap<long long, double, fph::meta::MixSeedHash<long long>>>
         ball_map;
 
     ApproxDistanceOracles(Graph* graph, long long k, bool sp, bool dbg);
     void updateRank();
     void chooseLandmarks();
     void printLandmarks();
-    void multiSourceDijkstra(const std::vector<int>& sources, long long level);
+    void multiSourceDijkstra(const std::vector<long long>& sources,
+                             long long level);
     void printFocii();
     void trimmedDijkstra(long long v, long long level);
     void printBalls();
     long long hashBalls();
     std::pair<long long, long long> preprocess(
-        const std::vector<std::vector<int>>& cust_land);
+        const std::vector<std::vector<long long>>& cust_land);
     double query(long long u, long long v);
 };
